Proposed solution for #152 (#166)
* Fix permission for get_all_users

* BUG FIX AND REFACTOR. Permissions on user routes have been corrected. Also, the functionality of user routings is now handled by a new service dedicated to user patching. This layout makes more sense and is more scalable.

* [REFACTOR]: Remove unnecessary print in services/user.py
