+++ routes/user_router.py
@@ -1,17 +1,17 @@
 from typing import Annotated
+from db_models import permission_model
 from fastapi import APIRouter, HTTPException, status
 from api_schemas.base_schema import BaseSchema
 from database import DB_dependency
 from db_models.user_model import User_DB
-from api_schemas.user_schemas import MeUpdate, UserRead
+from api_schemas.user_schemas import UpdateUserMember, UserUpdate, UserRead
+from services import user as user_service
 from user.permission import Permission
-from sqlalchemy.exc import DataError
-import re
 
 user_router = APIRouter()
 
 
-@user_router.get("/", response_model=list[UserRead])
+@user_router.get("/", response_model=list[UserRead], dependencies=[Permission.require("view", "User")])
 def get_all_users(db: DB_dependency):
     all_users = db.query(User_DB).all()
     return all_users
@@ -22,61 +22,15 @@ def get_me(user: Annotated[User_DB, Permission.base()]):
     return user
 
 
-def check_stil_id(s: str) -> bool:
-    if not len(s) == 10:
-        return False
-    pattern = r"^[a-z]{2}\d{4}[a-z]{2}-s$"
-    return bool(re.fullmatch(pattern, s))
+@user_router.patch("/update/me", response_model=UserRead)
+def update_self(data: UserUpdate, current_user: Annotated[User_DB, Permission.base()], db: DB_dependency):
+    return user_service.update_user(current_user.id, data, db)
+
+@user_router.patch("/update/{user_id}", response_model=UserRead, dependencies=[Permission.require("manage", "User")])
+def update_user(data: UserUpdate, user_id: int, db: DB_dependency):
+    return user_service.update_user(user_id, data, db)
 
 
-@user_router.patch("/me", response_model=UserRead)
-def update_me(data: MeUpdate, current_user: Annotated[User_DB, Permission.base()], db: DB_dependency):
-    # Since we edit user, look it up using "db" and not from permission so were are in same session
-    me = db.query(User_DB).filter_by(id=current_user.id).one()
-
-    if data.stil_id:
-        if not check_stil_id(data.stil_id):
-            raise HTTPException(400, detail="Invalid stil-id")
-        me.stil_id = data.stil_id
-
-    # not elegant, will have to find better wat for future update routes
-    if data.first_name:
-        me.first_name = data.first_name
-    if data.last_name:
-        me.last_name = data.last_name
-    if data.start_year:
-        me.start_year = data.start_year
-    if data.program:
-        me.program = data.program
-    if data.notifications is not None:
-        me.want_notifications = data.notifications
-
-    try:
-        db.commit()
-    except DataError:
-        db.rollback()
-        raise HTTPException(status.HTTP_400_BAD_REQUEST)
-
-    return current_user
-
-
-class UpdateUserMember(BaseSchema):
-    is_member: bool
-
-
-@user_router.patch("/member-status/{user_id}", dependencies=[Permission.require("manage", "User")])
-def update_user(user_id: int, data: UpdateUserMember, db: DB_dependency):
-    user = db.query(User_DB).filter_by(id=user_id).one_or_none()
-    if user is None:
-        raise HTTPException(status.HTTP_404_NOT_FOUND)
-
-    user.is_member = data.is_member
-
-    db.commit()
-    return user
-
-
-# TODO update your own stuff,
-
-
-# TODO delete routes
+@user_router.patch("/member-status/{user_id}", response_model=UserRead, dependencies=[Permission.require("manage", "User")])
+def update_user_status(user_id: int, data: UpdateUserMember, db: DB_dependency):
+    return user_service.update_user_status(user_id, data, db)
